requirements.txt
python-binance
python-dotenv
typer
